refactor: 모듈 디렉토리 구조 변경
