+++ apps/api-server/src/notification/application/dto/delete-notifications-result.dto.ts
