+++ apps/api-server/src/notification/notification.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Notification } from '@app/database';
-import { NotificationService } from './service/notification.service';
-import { NotificationController } from './controller/notification.controller';
+import { NotificationService } from './application/notification.service';
+import { NotificationController } from './api/notification.controller';
 
 @Module({
   imports: [MikroOrmModule.forFeature([Notification])],
