+++ apps/api-server/src/sns/application/kakao-sns.service.ts
@@ -1,10 +1,10 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { SnsService, SnsUserInfo } from './sns.service.interface';
-import { ErrorCode } from '../common/error-code';
+import { ErrorCode } from '../../common/error-code';
 
 @Injectable()
-export class NaverSnsService implements SnsService {
-  private readonly API_URL = 'https://openapi.naver.com/v1/nid/me';
+export class KakaoSnsService implements SnsService {
+  private readonly API_URL = 'https://kapi.kakao.com/v2/user/me';
 
   async getUserInfo(oAuthToken: string): Promise<SnsUserInfo> {
     try {
@@ -17,21 +17,20 @@ export class NaverSnsService implements SnsService {
       if (response.status === 401) {
         throw new UnauthorizedException({
           code: ErrorCode.INVALID_SNS_TOKEN,
-          message: 'Invalid Naver access token',
+          message: 'Invalid Kakao access token',
         });
       }
 
       const data = await response.json();
-      const userInfo = data.response;
 
       return {
-        id: userInfo.id,
+        id: data.id.toString(),
       };
     } catch (error) {
       if (error instanceof UnauthorizedException) {
         throw error;
       }
-      throw new Error(`Failed to fetch user info from Naver: ${error.message}`);
+      throw new Error(`Failed to fetch user info from Kakao: ${error.message}`);
     }
   }
 }
