+++ apps/api-server/src/sns/application/sns.service.interface.ts
