+++ apps/api-server/src/sns/sns.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { NaverSnsService } from './naver-sns.service';
-import { KakaoSnsService } from './kakao-sns.service';
-import { SnsServiceFactory } from './sns.service.factory';
+import { NaverSnsService } from './application/naver-sns.service';
+import { KakaoSnsService } from './application/kakao-sns.service';
+import { SnsServiceFactory } from './application/sns.service.factory';
 
 @Module({
   providers: [NaverSnsService, KakaoSnsService, SnsServiceFactory],
