+++ apps/api-server/src/user/api/dto/check-nickname-exists-response.dto.ts
