+++ apps/api-server/src/user/user.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { S3Module } from '@app/common/s3/s3.module';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { User, InterestRegion } from '@app/database';
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserController } from './api/user.controller';
+import { UserService } from './application/user.service';
 import { SnsModule } from '../sns/sns.module';
 import { AuthModule } from '../auth/auth.module';
